Extract and use `smartBlockWrap()`
